moved validators to validators.py
