+++ src/accounts/forms.py
@@ -5,14 +5,8 @@ from django.contrib.auth import password_validation
 from django.contrib.auth.forms import UserCreationForm
 from .models import CompanyProfile, StudentProfile
 from student.models import ProgramAndBranch
-from django.core.validators import RegexValidator
-from .validators import validate_year, validate_gpa, validate_phone, validate_domain, validate_url, validate_pincode
-
-
-def check_file_size(value):
-    limit = 3 * 1024 * 1024
-    if value.size > limit:
-        raise forms.ValidationError('File too large. Size should not exceed 3 MB.')
+from .validators import (
+    validate_year, validate_gpa, validate_phone, validate_domain, validate_url, validate_pincode, check_file_size, regex_validators)
 
 
 class ResumeForm(forms.ModelForm):
@@ -21,29 +15,6 @@ class ResumeForm(forms.ModelForm):
     class Meta:
         model = Resume
         fields = ('file', 'reference', 'student')
-
-
-validator_fn = [
-    RegexValidator(r'[A-Z]([A-Z]?)[0-9]{2}([A-Z]?)([A-Z]?)([A-Z]?)[0-9]{3}([0-9]?){4}',
-                   "Enter your Roll number(in correct "
-                   "format like eg. B17CS006 ). "
-                   "This will be used to login "),
-    RegexValidator(r'[A-Z]{2}[0-9]{2}[A-Z]([A-Z]?)([a-z]?)[A-Z][0-9]{3}',
-                   "Enter your Roll number(in correct "
-                   "format like eg. MT19VSS006 ). "
-                   "This will be used to login ")
-]
-
-
-def regex_validators(value):
-    err = None
-    for validator in validator_fn:
-        try:
-            validator(value)
-            return value
-        except forms.ValidationError as exc:
-            err = exc
-    raise err
 
 
 class StudentRegisterForm(UserCreationForm):
